Fade-out workaround
When you manually fade out a chase using a slider (or have the chase
dimmed and then stop the chase), the following will happen:
 - The chase will be stopped.
 - Its postRun will stop all steps/scenes currently running.
 - Its postRun will remove any attribute overrides from the
   steps/scenes.
 - The scene's postRun will initiate a fade out, but since the slider's
   override (which was forwarded by the chase) is no longer present, it
   initiates the fade from the original value.

To fix this, a function remembers the adjusted intensity when stopping
(can't do this in Scene::stop, since stop is not virtual) and a scene
uses that value when initiating the fade out.

Even with this change, moving a slider down would typically cause a
small fadeout (starting at a few percent, the value set by the slider
before the stop at value 0). This is fixed by setting the value to 0
before doing a stop (which still causes a fade out from 0 to 0
internally, but that's probably ok).